Combine group id pact into the single group-api pact file
